New tests and started implementing a grouping method
New tests are added for finding zones in the image.
The grouping methods are to be implemented with respect to movement.

For example for each action and for each point we should be able to
check if a given row is in a greedy zone or jumping zone for an action.
In order to do that we need to group the subareas in the zone with respect
to the action.
Each action then have subzones that are associated to it where, the
segmenting agent can simply jump.

The path cost can be calculated as the sum of the distances that is
necessary to arrive any of the goal states, plus the energy,
that is going to be associated with the path.